Move HE from app_common to app_opt. Update app_opt requirements management.
