+++ build_doc.sh
@@ -49,8 +49,7 @@ function clean_docs() {
 }
 
 function build_html_docs() {
-    # pip install -e ./
-    pip install -r requirements-doc.txt
+    pip install -e .[doc]
     sphinx-apidoc --module-first -f -o docs/apidocs/ nvflare "*poc" "*private"
     sphinx-build -b html docs docs/_build
 }
