+++ examples/advanced/tensorboard-streaming/README.md
@@ -1,16 +1,19 @@
 # Hello PyTorch with Tensorboard Streaming
 
-Example of using [NVIDIA FLARE](https://nvflare.readthedocs.io/en/main/index.html) to train an image classifier using federated averaging ([FedAvg]([FedAvg](https://arxiv.org/abs/1602.05629))) and [PyTorch](https://pytorch.org/) as the deep learning training framework. This example also highlights the TensorBoard streaming capability from the clients to the server.
+Example of using [NVIDIA FLARE](https://nvflare.readthedocs.io/en/main/index.html) to train an image classifier
+using federated averaging ([FedAvg]([FedAvg](https://arxiv.org/abs/1602.05629))) and [PyTorch](https://pytorch.org/)
+as the deep learning training framework.
+
+This example also highlights the TensorBoard streaming capability from the clients to the server.
 
 > **_NOTE:_** This example uses the [CIFAR-10](https://www.cs.toronto.edu/~kriz/cifar.html) dataset and will load its data within the trainer code.
 
-### 1. Install NVIDIA FLARE
+### 1. Install requirements
 
-Follow the [Installation](https://nvflare.readthedocs.io/en/main/getting_started.html) instructions.
 Install additional requirements:
 
 ```
-pip3 install -r requirements.txt
+python -m pip install -r requirements.txt
 ```
 
 ### 2. Run the experiment
