+++ examples/advanced/tensorboard-streaming/requirements.txt
@@ -1,3 +1,1 @@
-torch
-torchvision
-tensorboard
+nvflare [PT] >= 2.3.0
