--- requirements-dev.txt
@@ -1,11 +1,0 @@
--r requirements-doc.txt
-pydocstyle==6.1.1
-isort==5.10.1
-flake8==5.0.4
-black==22.10.0
-click==8.1.3
-pytest-xdist==3.0.2
-pytest-cov==4.0.0
-pandas==1.5.1
-# wait for approval
-# pylint!=2.13  # https://github.com/PyCQA/pylint/issues/5969
