+++ setup.cfg
@@ -35,9 +35,42 @@ install_requires =
 
 [options.extras_require]
 HE =
-    tenseal==0.3.0
+    tenseal==0.3.12
 PSI =
     openmined.psi==1.1.1
+PT =
+    torch
+    torchvision
+    tensorboard
+SKLEARN =
+    sklearn
+TRACKING =
+    mlflow
+app_opt =
+    %(HE)s
+    %(PSI)s
+    %(PT)s
+    %(SKLEARN)s
+    %(TRACKING)s
+doc =
+    sphinx>=4.1.1
+    sphinx_rtd_theme
+    recommonmark
+    sphinx-copybutton
+test =
+    %(app_opt)s
+    isort==5.10.1
+    flake8==5.0.4
+    black==22.10.0
+    click==8.1.3
+    pytest-xdist==3.0.2
+    pytest-cov==4.0.0
+    pandas==1.5.1
+dev =
+    %(doc)s
+    %(test)s
+all =
+    %(app_opt)s
 
 [options.entry_points]
 console_scripts =
