+++ tests/integration_test/data/apps/tb_streaming/config/config_fed_server.json
@@ -24,7 +24,7 @@
     },
     {
       "id": "tb_analytics_receive",
-      "path": "nvflare.app_opt.tracking.tensorboard.tb_receiver.TBAnalyticsReceiver",
+      "name": "TBAnalyticsReceiver",
       "args": {"events": ["fed.analytix_log_stats"]}
     }
   ],
